requirements.txt
python-telegram-bot==20.0
qrcode[pil]==7.3.1
python-dotenv==0.21.1
requests==2.28.2